+++ src/params.h
@@ -99,6 +99,9 @@ struct compression_params {
     std::string input_filename_1;
     std::string input_filename_2;
     std::string note;
+    uint32_t sequence_crc[2];
+    uint32_t quality_crc[2];
+    uint32_t id_crc[2];
   } read_info;
 };